initial refactor; current tests work
